feat(devops): Add GitHub linter (#125)
# Motivation
For security auditing, we use the zizmor linter for GitHub actions.

Zizmor has detected issues with eth current GitHub setup, so we start
off by adding it but not enforcing it. Once we have addressed the issues
identified by zizmor, we can enforce the lints.


# Changes
- Add zizmor to the list of dev tools.
- Add a  GitHub task (unenforced) that runs the zizmor linter.
- Add the zizmor linter to the overall lint command, but unenforced.

# Tests
The code is the same as for oisy-wallet, only with the lint command
renamed to match the conventions in this repo.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>